Clean up FormValidation module
